refactor(view): LadderFormat을 클래스 내부 enum으로 관리
